requirements/base.txt
Flask==3.0.0
pymongo==4.6.0
dnspython==2.4.2
pytz==2023.3.post1
pydantic==2.5.1
python-dotenv==1.0.0
email-validator==2.1.0.post1
typing_extensions==4.8.0
Werkzeug==3.0.1
annotated-types==0.6.0
blinker==1.7.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyppeteer==1.0.2
celery==5.3.5
redis==5.0.1
PyGithub==2.1.1
gunicorn==21.2.0